+++ packages/server/graphql/private/permissions.ts
@@ -8,6 +8,9 @@ const permissionMap: PermissionMap<Resolvers> = {
   },
   Query: {
     '*': isSuperUser
+  },
+  User: {
+    companyCluster: isSuperUser
   }
 }